rfctr: make @add_chunking_strategy a direct decorator
The `@add_chunking_strategy` decorator needs no parameters, but the
implementation adds an additional layer (function that returns a
function that returns a function) which adds unnecessary complexity.

Since we're in here anyway to extract the chunking dispatch, clean this
up.

Since this decorator will now cohere tightly with the rest of the
dispatch functionality, move it out of the initializer and into the
`dispatch` module.

Add unit tests for it while we're at it.
