+++ unstructured/partition/docx.py
@@ -168,7 +168,7 @@ def convert_and_partition_docx(
 
 @process_metadata()
 @add_metadata_with_filetype(FileType.DOCX)
-@add_chunking_strategy()
+@add_chunking_strategy
 def partition_docx(
     filename: Optional[str] = None,
     file: Optional[IO[bytes]] = None,
