+++ unstructured/partition/html.py
@@ -26,7 +26,7 @@ if TYPE_CHECKING:
 
 @process_metadata()
 @add_metadata_with_filetype(FileType.HTML)
-@add_chunking_strategy()
+@add_chunking_strategy
 def partition_html(
     filename: Optional[str] = None,
     file: Optional[IO[bytes]] = None,
