+++ unstructured/partition/pptx.py
@@ -47,7 +47,7 @@ DETECTION_ORIGIN = "pptx"
 
 @process_metadata()
 @add_metadata_with_filetype(FileType.PPTX)
-@add_chunking_strategy()
+@add_chunking_strategy
 def partition_pptx(
     filename: Optional[str] = None,
     file: Optional[IO[bytes]] = None,
