+++ unstructured/partition/rst.py
@@ -10,7 +10,7 @@ DETECTION_ORIGIN: str = "rst"
 
 @process_metadata()
 @add_metadata_with_filetype(FileType.RST)
-@add_chunking_strategy()
+@add_chunking_strategy
 def partition_rst(
     filename: Optional[str] = None,
     file: Optional[IO[bytes]] = None,
