+++ unstructured/partition/xml.py
@@ -81,7 +81,7 @@ def _get_leaf_elements(
 
 @process_metadata()
 @add_metadata_with_filetype(FileType.XML)
-@add_chunking_strategy()
+@add_chunking_strategy
 def partition_xml(
     filename: Optional[str] = None,
     file: Optional[Union[IO[bytes], SpooledTemporaryFile[bytes]]] = None,
